Rename muhenkan-switch to muhenkan-switch-core, muhenkan-switch-gui to muhenkan-switch
The GUI is the primary user-facing entry point, so it should own the
`muhenkan-switch` name. The CLI (called by kanata via cmd action) is
renamed to `muhenkan-switch-core`.

- Rename crate directories and update workspace members
- Update package/bin names in Cargo.toml files
- Update tauri.conf.json productName and identifier
- Update .kbd files: cmd muhenkan-switch → cmd muhenkan-switch-core
- Update all installer/uninstaller/update scripts for new binary names
- Update release.yml build steps and artifact names
- Update mise.toml dev tasks
- Update README.md and docs/design.md
