+++ Cargo.toml
@@ -3,5 +3,5 @@ resolver = "2"
 members = [
     "muhenkan-switch",
     "muhenkan-switch-config",
-    "muhenkan-switch-gui",
+    "muhenkan-switch-core",
 ]
