+++ kanata/muhenkan-macos.kbd
@@ -34,28 +34,28 @@
   mh (tap-hold 200 200 eisu (layer-while-held mh-layer))
 
   ;; 汎用ディスパッチ
-  dsp-1 (cmd muhenkan-switch dispatch 1)
-  dsp-2 (cmd muhenkan-switch dispatch 2)
-  dsp-3 (cmd muhenkan-switch dispatch 3)
-  dsp-4 (cmd muhenkan-switch dispatch 4)
-  dsp-5 (cmd muhenkan-switch dispatch 5)
-  dsp-q (cmd muhenkan-switch dispatch q)
-  dsp-r (cmd muhenkan-switch dispatch r)
-  dsp-t (cmd muhenkan-switch dispatch t)
-  dsp-g (cmd muhenkan-switch dispatch g)
-  dsp-a (cmd muhenkan-switch dispatch a)
-  dsp-w (cmd muhenkan-switch dispatch w)
-  dsp-e (cmd muhenkan-switch dispatch e)
-  dsp-s (cmd muhenkan-switch dispatch s)
-  dsp-d (cmd muhenkan-switch dispatch d)
-  dsp-f (cmd muhenkan-switch dispatch f)
+  dsp-1 (cmd muhenkan-switch-core dispatch 1)
+  dsp-2 (cmd muhenkan-switch-core dispatch 2)
+  dsp-3 (cmd muhenkan-switch-core dispatch 3)
+  dsp-4 (cmd muhenkan-switch-core dispatch 4)
+  dsp-5 (cmd muhenkan-switch-core dispatch 5)
+  dsp-q (cmd muhenkan-switch-core dispatch q)
+  dsp-r (cmd muhenkan-switch-core dispatch r)
+  dsp-t (cmd muhenkan-switch-core dispatch t)
+  dsp-g (cmd muhenkan-switch-core dispatch g)
+  dsp-a (cmd muhenkan-switch-core dispatch a)
+  dsp-w (cmd muhenkan-switch-core dispatch w)
+  dsp-e (cmd muhenkan-switch-core dispatch e)
+  dsp-s (cmd muhenkan-switch-core dispatch s)
+  dsp-d (cmd muhenkan-switch-core dispatch d)
+  dsp-f (cmd muhenkan-switch-core dispatch f)
 
   ;; 固定アクション
-  ts-v (cmd muhenkan-switch timestamp --action paste)
-  ts-c (cmd muhenkan-switch timestamp --action copy)
-  ts-x (cmd muhenkan-switch timestamp --action cut)
+  ts-v (cmd muhenkan-switch-core timestamp --action paste)
+  ts-c (cmd muhenkan-switch-core timestamp --action copy)
+  ts-x (cmd muhenkan-switch-core timestamp --action cut)
 
-  prtsc (cmd muhenkan-switch screenshot)
+  prtsc (cmd muhenkan-switch-core screenshot)
 )
 
 (deflayer default
