+++ muhenkan-switch-core/src/commands/open_folder.rs
