+++ muhenkan-switch-core/src/commands/screenshot.rs
