+++ muhenkan-switch-core/src/commands/search.rs
