+++ muhenkan-switch-core/src/commands/switch_app.rs
