--- muhenkan-switch-gui/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "muhenkan-switch-gui"
-version = "0.2.0"
-edition = "2021"
-description = "GUI settings manager for muhenkan-switch"
-license = "GPL-2.0-only"
-
-[dependencies]
-muhenkan-switch-config = { path = "../muhenkan-switch-config" }
-anyhow = "1"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tauri = { version = "2", features = ["tray-icon"] }
-tauri-plugin-shell = "2"
-tauri-plugin-dialog = "2"
-tauri-plugin-autostart = "2"
-tauri-plugin-store = "2"
-shared_child = "1"
-chrono = "0.4"
-open = "5"
-
-[target.'cfg(windows)'.dependencies]
-windows = { version = "0.61", features = [
-    "Win32_Security",
-    "Win32_System_Diagnostics_ToolHelp",
-    "Win32_System_JobObjects",
-    "Win32_System_Threading",
-] }
-
-[build-dependencies]
-tauri-build = { version = "2", features = [] }
