+++ muhenkan-switch/Cargo.toml
@@ -2,35 +2,30 @@
 name = "muhenkan-switch"
 version = "0.2.0"
 edition = "2021"
-description = "Companion tool for muhenkan-switch-rs: cross-platform shortcut actions"
+description = "GUI settings manager for muhenkan-switch"
 license = "GPL-2.0-only"
-repository = "https://github.com/kimushun1101/muhenkan-switch-rs"
-
-[[bin]]
-name = "muhenkan-switch"
-path = "src/main.rs"
 
 [dependencies]
 muhenkan-switch-config = { path = "../muhenkan-switch-config" }
-clap = { version = "4", features = ["derive"] }
-arboard = { version = "3", features = ["wayland-data-control"] }
-webbrowser = "1"
-open = "5"
-chrono = "0.4"
-urlencoding = "2"
 anyhow = "1"
-dirs = "6"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tauri = { version = "2", features = ["tray-icon"] }
+tauri-plugin-shell = "2"
+tauri-plugin-dialog = "2"
+tauri-plugin-autostart = "2"
+tauri-plugin-store = "2"
+shared_child = "1"
+chrono = "0.4"
+open = "5"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.61", features = [
-    "Win32_System_Com",
-    "Win32_System_LibraryLoader",
+    "Win32_Security",
     "Win32_System_Diagnostics_ToolHelp",
-    "Win32_System_Ole",
+    "Win32_System_JobObjects",
     "Win32_System_Threading",
-    "Win32_System_Variant",
-    "Win32_UI_Input_KeyboardAndMouse",
-    "Win32_UI_Shell",
-    "Win32_Graphics_Gdi",
-    "Win32_UI_WindowsAndMessaging",
 ] }
+
+[build-dependencies]
+tauri-build = { version = "2", features = [] }
