+++ muhenkan-switch/gen/schemas/capabilities.json
