+++ muhenkan-switch/gen/schemas/desktop-schema.json
