+++ muhenkan-switch/icons/icon.icns
