+++ muhenkan-switch/src/tray.rs
