+++ muhenkan-switch/tauri.conf.json
@@ -1,8 +1,8 @@
 {
   "$schema": "https://raw.githubusercontent.com/tauri-apps/tauri/dev/crates/tauri-cli/schema.json",
-  "productName": "muhenkan-switch-gui",
+  "productName": "muhenkan-switch",
   "version": "0.2.0",
-  "identifier": "com.muhenkan-switch.gui",
+  "identifier": "com.muhenkan-switch",
   "build": {
     "frontendDist": "./frontend"
   },
