+++ scripts/uninstall.sh
@@ -6,7 +6,7 @@ set -euo pipefail
 INSTALL_DIR="$HOME/.local/share/muhenkan-switch-rs"
 BIN_DIR="$HOME/.local/bin"
 SERVICE_FILE="$HOME/.config/systemd/user/kanata.service"
-AUTOSTART_FILE="$HOME/.config/autostart/muhenkan-switch-gui.desktop"
+AUTOSTART_FILE="$HOME/.config/autostart/muhenkan-switch.desktop"
 
 echo ""
 echo "=== muhenkan-switch-rs アンインストーラー (Linux) ==="
@@ -20,7 +20,7 @@ fi
 
 echo "以下を削除します:"
 echo "  - インストールディレクトリ: $INSTALL_DIR"
-echo "  - シンボリックリンク: $BIN_DIR/muhenkan-switch, $BIN_DIR/muhenkan-switch-gui, $BIN_DIR/kanata_cmd_allowed"
+echo "  - シンボリックリンク: $BIN_DIR/muhenkan-switch, $BIN_DIR/muhenkan-switch-core, $BIN_DIR/kanata_cmd_allowed"
 if [ -f "$SERVICE_FILE" ]; then
     echo "  - systemd サービス: kanata.service"
 fi
@@ -75,6 +75,7 @@ remove_symlink() {
 }
 
 remove_symlink "muhenkan-switch"
+remove_symlink "muhenkan-switch-core"
 remove_symlink "muhenkan-switch-gui"
 remove_symlink "kanata_cmd_allowed"
